+++ src/Resources/contao/languages/de/tl_layout.php
@@ -7,13 +7,12 @@ $lang = &$GLOBALS['TL_LANG']['tl_layout'];
  */
 $lang['addEncore'][0]                        = 'Webpack Encore aktivieren';
 $lang['addEncore'][1]                        = 'Wählen Sie diese Option, wenn Sie die Encore-Unterstützung aktivieren möchten.';
-$lang['addEncoreBabelPolyfill'][0]           = 'babel-polyfill hinzufügen (für IE <= 11)';
-$lang['addEncoreBabelPolyfill'][1]           =
-    'Wählen Sie diese Option, wenn der Internet Explorer in Version <= 11 unterstützt werden soll. In modernen Browsern (auch Edge) wird das Polyfill, das u.a. Promises bereitstellt, in der Regel nicht benötigt.';
+$lang['addEncoreBabelPolyfill'][0]           = 'Veraltet: babel-polyfill hinzufügen';
+$lang['addEncoreBabelPolyfill'][1]           = 'Diese Option ist veraltet und sollte nicht mehr verwendet werden. Das Babel-Polyfill kann einfach als aktives Entry hinzugefügt werden. ';
 $lang['encoreBabelPolyfillEntryName'][0]     = 'babel-polyfill-Entry-Name';
 $lang['encoreBabelPolyfillEntryName'][1]     =
     'Geben Sie hier den Namen des babel-polyfill-Entry ein, wie er Encore.addEntry() als erster Parameter übergeben wird.';
-$lang['encoreEntries'][0]                    = 'Aktive Entrys';
+$lang['encoreEntries'][0]                    = 'Aktive Entries';
 $lang['encoreEntries'][1]                    =
     'Legen Sie hier fest, welche webpack-Entrys auf welchen Seiten gerendert werden sollen. Sie können diese Festlegungen auf eventuellen Unterseiten mit Seitenvererbung überschreiben.';
 $lang['encoreEntries_entry'][0]              = 'Entrys';