+++ src/Resources/contao/languages/de/tl_page.php
@@ -5,16 +5,11 @@ $lang = &$GLOBALS['TL_LANG']['tl_page'];
 /**
  * Fields
  */
-    'Geben Sie hier den Namen des babel-polyfill-Entry ein, wie er Encore.addEntry() als erster Parameter übergeben wird.';
-$lang['encoreEntries'][0]                    = 'Aktive Entrys';
+$lang['encoreEntries'][0]                    = 'Aktive Entries';
 $lang['encoreEntries'][1]                    =
     'Legen Sie hier fest, welche webpack-Entrys auf welchen Seiten gerendert werden sollen. Sie können diese Festlegungen auf eventuellen Unterseiten mit Seitenvererbung überschreiben.';
-$lang['encoreEntries_entry'][0]              = 'Entrys';
+$lang['encoreEntries_entry'][0]              = 'Entries';
 $lang['encoreEntries_active'][0]             = 'Aktiv';
-$lang['encoreStylesheetsImportsTemplate'][0] = 'Alternatives Stylesheets Import-Template';
-$lang['encoreStylesheetsImportsTemplate'][1] = 'Wählen Sie hier bei Bedarf ein alternatives Import-Template für Stylesheets aus.';
-$lang['encoreScriptsImportsTemplate'][0]     = 'Alternatives Javascript Import-Template';
-$lang['encoreScriptsImportsTemplate'][1]     = 'Wählen Sie hier bei Bedarf ein alternatives Import-Template für Javascript aus.';
 
 /**
  * Legends